adjust getSelectionTextFromEditor method location
